+++ src/curios_comic_crawler/__init__.py
@@ -1,5 +1,7 @@
 """CuriosComicCrawler: download and locally upscale Sequential Art comic pages."""
 
+from importlib.metadata import version
+
 __author__ = 'Luraminaki'
 __email__ = ''
-__version__ = '0.4.0'
+__version__ = version('curios-comic-crawler')